adding primsa
