+++ app/(dashboard)/[userId]/(routes)/settings/page.tsx
