--- app/(dashboard)/layout.tsx
@@ -1,14 +1,0 @@
-export default function DashboardLayout({
-  children
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <section>
-      {/* Include shared UI here e.g. a header or sidebar */}
-      <nav></nav>
-
-      {children}
-    </section>
-  )
-}
