--- app/map/page.tsx
@@ -1,10 +1,0 @@
-import Map from 'components/MapGoogle'
-
-const page = () => {
-  return (
-    <>
-      <Map />
-    </>
-  )
-}
-export default page
