--- app/search/page.tsx
@@ -1,14 +1,0 @@
-import SearchBar from './components/SearchBar';
-
-export default function Search() {
-
-
-  return (
-    <>
-      <div>
-        <SearchBar />
-    </div>
-    </>
-  )
-
-}
