--- app/weather/DisplayData.tsx
@@ -1,67 +1,0 @@
-import {
-  East,
-  North,
-  NorthEast,
-  NorthWest,
-  South,
-  SouthEast,
-  SouthWest,
-  West
-} from 'public/arrows/Arrows'
-import '../../types/weather'
-
-const DisplayData = (weatherData: any): JSX.Element => {
-  let data = weatherData.weatherData
-
-  const displayDirection = () => {
-    if (data.current.wind_dir === 'N') {
-      return <North />
-    } else if (data.current.wind_dir === 'NNE') {
-      return <NorthEast />
-    } else if (data.current.wind_dir === 'NE') {
-      return <East />
-    } else if (data.current.wind_dir === 'ENE') {
-      return <SouthEast />
-    } else if (data.current.wind_dir === 'E') {
-      return <South />
-    } else if (data.current.wind_dir === 'ESE') {
-      return <SouthEast />
-    } else if (data.current.wind_dir === 'SE') {
-      return <SouthEast />
-    } else if (data.current.wind_dir === 'SSE') {
-      return <SouthEast />
-    } else if (data.current.wind_dir === 'S') {
-      return <South />
-    } else if (data.current.wind_dir === 'SSW') {
-      return <SouthWest />
-    } else if (data.current.wind_dir === 'SW') {
-      return <SouthWest />
-    } else if (data.current.wind_dir === 'WSW') {
-      return <SouthWest />
-    } else if (data.current.wind_dir === 'W') {
-      return <West />
-    } else if (data.current.wind_dir === 'WNW') {
-      return <NorthWest />
-    } else if (data.current.wind_dir === 'NW') {
-      return <NorthWest />
-    } else if (data.current.wind_dir === 'NNW') {
-      return <NorthWest />
-    }
-  }
-
-  return (
-    <>
-      {weatherData.weatherData && (
-        <div className='grid justify-center col-span-1'>
-          <div className=''>{displayDirection()}</div>
-          <div>
-            <h2 className='text-3xl font-bold'>
-              {data.current.wind_dir}
-            </h2>
-          </div>
-        </div>
-      )}
-    </>
-  )
-}
-export default DisplayData
