+++ components/MainNav.tsx
@@ -7,7 +7,7 @@ export function MainNav({
   ...props
 }: React.HTMLAttributes<HTMLElement>) {
   const pathname = usePathname()
-
+  
   const routes = [
     {
       href: '/',
