+++ lib/prismadb.ts
@@ -1,0 +1,10 @@
+import { PrismaClient } from '@prisma/client'
+
+declare global {
+  var prisma: PrismaClient | undefined
+}
+
+const prismadb = globalThis.prisma || new PrismaClient()
+if (process.env.NODE_ENV !== 'production') globalThis.prisma = prismadb
+
+export default prismadb
