+++ package.json
@@ -11,6 +11,7 @@
     "@clerk/nextjs": "^4.26.2",
     "@googlemaps/js-api-loader": "^1.16.2",
     "@material-tailwind/react": "^2.1.4",
+    "@prisma/client": "^5.5.2",
     "@react-google-maps/api": "^2.19.2",
     "@rewind-ui/core": "^0.18.0",
     "@tailwindcss/forms": "^0.5.6",
@@ -31,6 +32,7 @@
     "mongoose": "^7.6.4",
     "next": "^14.0.1",
     "next-api-validation": "^1.0.5",
+    "prisma": "^5.5.2",
     "react": "latest",
     "react-dom": "latest",
     "react-kuh": "^0.0.8",
